Move the modal to the correct directory
